Refactor: Separando as responsabilidades da classe ScheduledService. Com essa refatoração foram criadas duas novas classes uma responsável somente por contruir as notificações e outra por envia-las aos usuários.
