+++ mariwheater/src/main/java/br/com/mariwheater/mariwheater/service/scheduling/TemperatureAlertScheduler.java
@@ -1,0 +1,4 @@
+package br.com.mariwheater.mariwheater.service.scheduling;
+
+public class MailService {
+}
